By default we enable background mode when RWF mode is enabled.
This makes us automatically refresh the Guild News every 30 seconds as long we are out of combat.
While the RWF frame is closed/hidden, during this mode we will display a small RIO button with our logo and a counter indicating loot ready to be written to disk.
Left-click to ReloadUI and save the data to the SV file. Right-click always displays the RWF frame.
The background mode makes it so after reloads the internals still keep monitoring the Guild News.
Currently we haven't added a toggle for the mode, it's on by default, possibly acceptable as the user does have to enable RWF mode to begin with?