Created `ns.MARKUP_ICONS` as a layer between markup icons that rely on atlas, so that we can support multiple clients, some that might not have those atlas available.
Used the older logic for handling Cataclysm WhoFrame. This might be needed for the Guild Roster (have no access to a guild to test this out right now).

Updated the docs for `ExecuteWidgetOnEnterSafely` to take in a Frame or a ScriptRegion so we could drop the disable diagnostic comment everywhere we wanted to pass it the `GetMouseFocus()` value.

Enabled the community tooltip as the code inside will ensure that the frames exist anyway, and the system does exist on Cataclysm.

Updated the shortcuts module to create `TooltipHelpText` variable that will be used directly as the help text when the user hovers the minimap button.

This variant has fallback in case the client doesn't have the icon available, to instead show the icons as text.