Created DropDownUtil to help migration from the old DropDown system to the new Menu system.
The DropDownUtil will assist detecting which codebase to use (what is available). It also contains the specific code to these two systems to create, initialize, open/close, and state detection logic.

The code that actually implements the UI is kept in the modules needing this, but, they use DropDownUtil to create and to keep track of which API to use/call.

Currently the Settings frame and the Replay frame both use dropdowns and both should be migrated to be able to use both systems.

Need to implement OpenMenu, IsMenuOpen and CloseMenu for 11.0.