Noticed Cataclysm had recruitment files generated, but not loaded by the Cata.toc files, so ran the toc.ps1 and added them for loading.
Updated the TOC versions for the various games to the current latest, since I noticed some were slightly behind.

Moved some keystone timing code into utilities, along with the special hotfix that adds 90 seconds to all times at 7 and above, so we can easily disable this when the time arrives, or adjust it to new logic as needed.

Updated AppendRecentRunsWithCharacter to rely on the profile being shown, since the keystone profile wasn't being used, this way we can also just read the name and realm that we know is correct, directly from the profile.

Minor changes, added docs for typing checks, overall the code does what it did before.