Added test warband icon to the icons file.
Added missing doc definitions related to warband fields.

Updated `GetSortedDungeonsTooltipText` so it supports generating data for the character and the warband.

Updated the profile rendering code to use the new methods to generate data for the character and the warband, then properly add the column and padding.

This was based on the old tyrannical/fortified code which did the same column behavior, just that we only display the warband if it exists in the data.

Created `ns.PROFILE_TOOLTIP_COLUMN_TEXTURE` to keep the column icons used in the profile tooltip.

Fixed minor copy JSON issue, trying to show when it can't via the slash command. Also commented out some taint problematic code, no longer needed.

Moved some stray static popup methods into the `StaticPopupUtil` collection.

Replaced a lot of string concatinations with format variants for clearer readability.