improving warband score display (#336)
Added missing doc definitions related to warband fields.

Updated `GetSortedDungeonsTooltipText` so it supports generating data for the character and the warband.

Updated the profile rendering code to use the new methods to generate data for the character and the warband, then properly add the column and padding.

This was based on the old tyrannical/fortified code which did the same column behavior, just that we only display the warband if it exists in the data.

Created `ns.PROFILE_TOOLTIP_COLUMN_TEXTURE` to keep the column icons used in the profile tooltip.

Fixed minor copy JSON issue, trying to show when it can't via the slash command. Also commented out some taint problematic code, no longer needed.

Moved some stray static popup methods into the `StaticPopupUtil` collection.

Replaced a lot of string concatinations with format variants for clearer readability.

Resized the warband icons down to 70% so they aren't too large, and made the icon just naturally fill the available space in the tooltip line.

Set the x-offset to the warband icon to 4 pixels.

Vertical offset on warband icons to align with font baseline better

Add a space-character to fix the issue with alignment and the digit 1 in the text on the right-side.

---------

Co-authored-by: Alex Pedersen <vladix@gmail.com>